# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api_handlers.albumResponse:
    properties:
      artist_id:
        type: string
      cover_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      release_date:
        type: string
      title:
        type: string
    type: object
  internal_api_handlers.artistResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  internal_api_handlers.createAlbumRequest:
    properties:
      artist_id:
        type: string
      release_date:
        description: ReleaseDate is optional and, being a DATE, carries no time or
          zone.
        type: string
      title:
        type: string
    type: object
  internal_api_handlers.createArtistRequest:
    properties:
      bio:
        type: string
      name:
        type: string
    type: object
  internal_api_handlers.createPlaylistRequest:
    properties:
      description:
        type: string
      title:
        type: string
      track_ids:
        items:
          type: string
        type: array
    type: object
...
  internal_api_handlers.playlistResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      track_ids:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Music library management and streaming API
  title: Mydal API
  version: "1.0"
paths:
  /albums:
    post:
      consumes:
      - application/json
      description: Add a new album to the database
      parameters:
      - description: Album payload
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.createAlbumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.albumResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new album
      tags:
      - albums
  /albums/{id}:
    delete:
      description: Remove an album from the database by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an album
      tags:
      - albums
    get:
      description: Retrieve a single album by its unique identifier
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.albumResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get album by ID
      tags:
      - albums
  /artists:
    post:
      consumes:
      - application/json
      description: Add a new artist to the database
      parameters:
      - description: Artist payload
        in: body
        name: artist
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.createArtistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.artistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new artist
      tags:
      - artists
  /artists/{id}:
    delete:
      description: Remove an artist from the database by ID
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an artist
      tags:
      - artists
    get:
      description: Retrieve a single artist by their unique identifier
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.artistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get artist by ID
      tags:
      - artists
  /playlists:
    post:
      consumes:
      - application/json
      description: Add a new playlist to the database
      parameters:
      - description: Playlist payload
        in: body
        name: playlist
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.createPlaylistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.playlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new playlist
      tags:
      - playlists
  /playlists/{id}:
    delete:
      description: Remove a playlist from the database by ID
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a playlist
      tags:
      - playlists
    get:
      description: Retrieve a single playlist by its unique identifier
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.playlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get playlist by ID
      tags:
      - playlists
  /playlists/{id}/tracks/{trackId}:
    delete:
      description: Remove a track from an existing playlist
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove track from playlist
      tags:
      - playlists
    put:
      description: Add an existing track to an existing playlist
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add track to playlist
      tags:
      - playlists
...
swagger: "2.0"